first step of merge
